requirements.txt
streamlit
python-dotenv
requests
mcp
langchain
langchain-google-genai
